Added a 'domainname' template tag.
Also, something is horribly buggy with either restructuredText, or
admindocs' parsing of it. The following docstring:

    """
    Return the domain name of the given URL. For example::

        {{ 'http://example.org/stuff'|domainname }}

    would display ``example.org``.
    """

Raises the error "Literal block expected; none found." on line 2
of the docstring (the same error is also raised throughout crispy-forms'
own docs). No amount of fiddling with things seems to fix it, except for
changing the docstring to the following:

    """
    Return the domain name of the given URL.

    For example::

        {{ 'http://example.org/stuff'|domainname }}

    would display ``example.org``.
    """

Debugging this was quite frustrating.