Made 'hive_member_since' an optional field.
This will allow us to distinguish between organizations that are actually
Hive members, and organizations that are "friends of Hive", e.g. partners.

Also changed the language in various places to not refer to 'hive member
organization' unless absolutely necessary: when possible, we'll use the
term 'hive organization' instead, to include both organizations that are
partners and members.