Added a search results page.
The typeahead autocomplete only completes characters in an organization's
name and a person's full name. Now a final result called 'Search the
website for "<your query>"' is on there as well, and it links to a
search results page that searches the database more
thoroughly--specifically,through organization mission, user bio and
title.