Added a 'slug' field to city.
I also discovered that using call_command() to import fixtures in a
migration, as the South docs suggest, doesn't work when the model
loading the fixtures changes, since it's using the latest version of
the model rather than South's serialized orm object. So I had to
create the fixture in code instead, which is honestly more readable
anyways.