Added city-scoped searches at `/<city-slug>/`.
This is kind of a weird commit that leaves several things in flux.

I'm trying to see if I can make this project something that can be
deployed *either* as a city-specific Hive directory *or* an umbrella site
that houses the directories of many hives, depending on whether or not
the current site (identified by settings.SITE_ID) is associated with a
City model or not.

Currently, a site is potentially both of these things at once, which
is definitely odd.

One of the things I don't like about this commit is that it removes the
use of reverse urls for /find.json and /search. Right now it looks like
this might need to be fixed at least partly through the use of a custom
template tag.