Add MULTI_CITY setting, city_url template tag, city_reverse().
This is a slight improvement over the last commit, which introduces the
concept of a "city URL", that is, any URL whose view name begins with
"city_". On multi-city configurations, a city URL will begin with the
slug of a city, e.g. `/nyc/`, while on non multi-city installs they
will be unprefixed. The accompanying template tag and URL resolver
take care of figuring out how the site is configured and providing the
proper URL to the caller.